Move all logic required in Spree::CheckoutController to CheckoutController
