+++ app/controllers/checkout_controller.rb
@@ -1,7 +1,13 @@
 require 'open_food_network/address_finder'
 
-class CheckoutController < Spree::CheckoutController
+class CheckoutController < Spree::StoreController
   layout 'darkswarm'
+
+  include CheckoutHelper
+  include OrderCyclesHelper
+  include EnterprisesHelper
+
+  ssl_required
 
   # We need pessimistic locking to avoid race conditions.
   # Otherwise we fail on duplicate indexes or end up with negative stock.
@@ -12,10 +18,19 @@ class CheckoutController < Spree::CheckoutController
   prepend_before_filter :require_order_cycle
   prepend_before_filter :require_distributor_chosen
 
+  before_filter :load_order
+
+  before_filter :ensure_order_not_completed
+  before_filter :ensure_checkout_allowed
+  before_filter :ensure_sufficient_stock_lines
+
+  before_filter :associate_user
+  before_filter :check_authorization
   before_filter :enable_embedded_shopfront
 
-  include OrderCyclesHelper
-  include EnterprisesHelper
+  helper 'spree/orders'
+
+  rescue_from Spree::Core::GatewayError, with: :rescue_from_spree_gateway_error
 
   def edit
     # This is only required because of spree_paypal_express. If we implement
@@ -86,6 +101,25 @@ class CheckoutController < Spree::CheckoutController
   end
 
   private
+
+  def check_authorization
+    authorize!(:edit, current_order, session[:access_token])
+  end
+
+  def ensure_checkout_allowed
+    redirect_to main_app.cart_path unless @order.checkout_allowed?
+  end
+
+  def ensure_order_not_completed
+    redirect_to main_app.cart_path if @order.completed?
+  end
+
+  def ensure_sufficient_stock_lines
+    if @order.insufficient_stock_lines.present?
+      flash[:error] = Spree.t(:inventory_error_flash_for_insufficient_quantity)
+      redirect_to main_app.cart_path
+    end
+  end
 
   def set_default_bill_address
     if params[:order][:default_bill_address]
@@ -185,6 +219,11 @@ class CheckoutController < Spree::CheckoutController
     setup_for_current_state
   end
 
+  def setup_for_current_state
+    method_name = :"before_#{@order.state}"
+    send(method_name) if respond_to?(method_name, true)
+  end
+
   def before_address
     associate_user
 
@@ -192,6 +231,17 @@ class CheckoutController < Spree::CheckoutController
 
     @order.bill_address = finder.bill_address
     @order.ship_address = finder.ship_address
+  end
+
+  def before_delivery
+    return if params[:order].present?
+
+    packages = @order.shipments.map(&:to_package)
+    @differentiator = Spree::Stock::Differentiator.new(@order, packages)
+  end
+
+  def before_payment
+    current_order.payments.destroy_all if request.put?
   end
 
   def valid_order_line_items?
